requirements.txt
aiohttp
bs4
discord
ffmpeg
numpy
pandas
python-dotenv
pytube
requests-html
tweepy